Avoid undefined right-shifting in jcdctmgr.c's quantize().
Since compute_reciprocal() with a divisor argument of 1 results in a
negative shift entry in the DCT table, we might as well use that to
identify that the divisor was 1 originally.

In that specific case, avoid doing the multiplication and shifting magic
in quantize(), and simply copy the temp coefficient to coef_blocks[]
unchanged.