fix crash when compiling with Clang for AArch64
This patch fixes https://github.com/android-ndk/ndk/issues/110
The crash seemed to come from a change in LLVM and was reported to the LLVM
bugzilla https://llvm.org/bugs/show_bug.cgi?id=28393

Upon investigation, LLVM optimized away the zeroing of the upper bits of the x2
register before calling color_convert.  The convention is for the callee to do
any truncation needed.  The patch tries to use the Wn registers whenever
possible, and otherwise use a zero-extend instruction to avoid using the
random information in the upper 32 bits of the 64 bit registers.