Fix AArch64 ABI conformance issue in SIMD code
In the AArch64 ABI, as in many others, it's forbidden to read/store data below
the stack pointer.

Some SIMD functions were doing just that (stack pointer misuse) when trying to
preserve callee-saved registers which resulted in those being restored to wrong
values in specific conditions.

This patch fixes that behavior and callee-saved registers are now stored above
the stack pointer throughout the function call.
It also removes unnecessary register saving in this specific ABI or makes use
of unused scratch regiters instead of callee-saved registers.

Closes #97. Refer also to https://bugzilla.redhat.com/show_bug.cgi?id=1368569