AppVeyor: Use 1.5.x buildscripts branch
The buildscripts repository is maintained such that its branch structure
matches that of the libjpeg-turbo repository, so we need to clone that
repository using the same branch name as the base branch of the build.
This allows us to continue using the autotools-based official build
scripts that were used for 1.5.x

This was implemented in the 2.0.x code base as part of the build system
overhaul (6abd39160c5a3762e9ebe024e75407665093e715) but was never
back-ported to 1.5.x.