Provide memory manager overriding api
This provides a new interface for creating compress and decompress
streams with an already initialized memory manager. By adding a new
function this doesn't break compatability with old clients. The new api
assumes that either mem has been initialized and is ready to go or that
it is NULL and will be initialized as normal.

Without this it is problematic to actually fully override the default
memory manager. The main issue is that the memory manager is that the
jpeg_create_* functions initialize the default memory manager and then
use it to allocate some items before returning. If the memory manager is
replaced after this the original memory manager will never have the
destruct called. If the destruct is called manually it will free memory
that is still in use by structures in cinfo.