Use builtin CMake module GNUInstallDirs instead of the local copy.
This fixes LIBDIR path composition with CMake 3.13.2. The local copy
would append the absolute LIBDIR to the PREFIX path instead of using
the relative path, if CMAKE_INSTALL_LIBDIR is specified by user.

Since this change requires GNUInstallDirs to be part of CMake
distribution, bumped minimum CMake version to 3.0.2 (the oldest one
that has documentation for this module).