Fix unused variable & functions in ARM SIMD
When simd/arm/jsimd.c is compiled with __ARM_NEON__ defined there
is a bufsize variable which is defined but never used.

Additionally, parse_proc_cpuinfo() is never used. That function is the
only one which calls check_feature() so it is also never used.

This commit puts all of these under the appropriate #if defined() so
there are no longer unused variables and functions.