i386/x86_64: make jsimd_idct_islow/_ifast/_float memory sanitizer friendly
This fixes an issue when analyzing an oss-fuzz on the libtiff library when
a file using JPEG compression is used.
https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=7547

When building both libjpeg-turbo and libtiff with -fsanitize=memory (MSAN),
that one fails to detect when assembly code sets buffer. Consequently, it
reports false positive erros in libtiff code.

Add a init_output_buf() macro/function that evaluates to no-op when MSAN is
not enabled, and when it is enabled, 0-initialize output_buf before going to
the assembly code.