Make what was JERR_BAD_STATE errors abort, as they are programming errors
As far as I can tell, any time a JERR_BAD_STATE is emitted to
error_exit(), it means that the libjpeg-turbo API has been called in
an incorrect order.  This is a programming error and the caller cannot
recover from it.  Thus, we'll make the library abort instead of
transferring control back to the caller.