Add 32-byte memory alignment check for Arm NEON
ALIGN_SIZE is defined to be 32 bytes in order to accommodate AVX2's
alignment requirements. Arm NEON only requires 16-byte alignment but
software optimizations are possible if 32-byte alignment is assumed
for memory allocated by jmemmgr.c.

This patch introduces a compile-time check to make sure that
ALIGN_SIZE is a multiple of 32 bytes for Arm NEON systems.

Optimizations relying on this assumption for Arm NEON will follow
this commit.

Change-Id: Ia49e7745ffa126e8ca98a7cc856adc3577117b42