Revert "Fix MIPS DSPr2 build when using soft float ABI"
This reverts commit 3bef88f6ecc08a5d3babffbf2c12175c2c1985b0.

The original goal of this commit was to fix compilation when using soft
float. However even though it compiles, the results that are produced are
totally bogus.

The MIPS ASM seems to also be marginally slower on the hardware that I
have tested it on. All of it without hardware floating-point support.