Add compiler-independent alignment macro
Some variables and structures need memory alignment and the syntax
for declaring it differs when using MSVC compared to GCC and Clang.

This commit adds a compiler-independent alignment macro that will be
used provide memory alignment for constant pools - it will be used in
future commits adding Arm Neon intrinsics code.

Change-Id: If1e6e203f1ebfb69da06216b3706c42933bc35a3