Use CLZ instructions for all Arm builds
CLZ instructions are currently used in the Huffman encoding process
on all GNUC-compiliant Arm platforms. This change adds support for
Microsoft's count-leading-zeros builtins on the Windows on Arm
platform.

As well as making the scalar Huffman encoding paths faster, this
change means the jpeg_nbits_table is no longer included for Windows
on Arm builds of libjpeg-turbo - saving 128KB of memory.

Change-Id: Iee992e74cbfcd46cd74aeb449a038e8c6ace4efb