Fix Neon intrinsics Huffman encoding for WoA/clang-cl
The GNU builtin function __builtin_clzl() takes an argument of type
unsigned long - this is 8 bytes on all systems other than Windows,
where it is 4 bytes. This caused the Neon intrinsics implementation
of Huffman encoding to fail when compiled using clang-cl for the
Windows on Arm platform.

Change all instances of __builtin_clzl() to __builtin_clzll() - which
takes an argument of type unsigned long long, guaranteed to be 8
bytes on all systems.

Change-Id: I78211c1d5c832ca3deda1adf439d0f77d4e689c6