jchuff.h: fix undefined behaviour in the FLUSH macro on ARM 32-bit
This commits fixes an undefined behaviour in the FLUSH macro on ARM
32-bit.

According to the C standard [1] (6.3.2.3 Pointers), and this llvm issue
[2], the following part of the FLUSH macro:

    *((uint32_t *)buffer) = BUILTIN_BSWAP32(put_buffer);
    buffer += 4;

is undefined behaviour and gets compiled by various clang versions
(using -O2 -mthumb or -Oz -mthumb) to:

    rev     r0, r0
    stmia   r1!, {r0}

Causing a crash (sigbus: illegal alignment) when the destination buffer
is not aligned. This notably happens when compiling libjpeg-turbo for
Android ARM 32-bit Thumb (using NDK r21e, r22b, ...).

This fix (using memcpy) compiles to:

    rev     r0, r0
    str.w   r0, [r1], #4

is not subject to alignment issues and is the way to go according to
[2].

[1]: http://www.open-std.org/jtc1/sc22/WG14/www/docs/n1256.pdf
[2]: https://bugs.llvm.org/show_bug.cgi?id=50785