Cast pointer arguments to PAD() to uintptr_t instead of size_t
The C standard does not guarantee that size_t is large enough to hold a
pointer value (e.g. 32-bit size_t and 64-bit pointers), so this cast can
strip parts of the pointer value. Casting to uintptr_t avoids this problem
since (u)intptr_t is defined as the integer type that can hold a pointer.

This change allows libjpeg-turbo to be compiled for CHERI-enabled
architectures (e.g. CHERI-RISC-V or Arm's Morello), since pointers are
twice the size of size_t (128 bits for Morello and RV64) and this cast
via size_t creates a non-dereferenceable pointer.
This is flagged by the following compiler warning:
```
warning: cast from provenance-free integer type to pointer type will give pointer that can not be dereferenced [-Werror,-Wcheri-capability-misuse]
  cvalue = values = (JCOEF *)PAD((size_t)values_unaligned, 16);
```
Not fixing this warning will result in a run-time crash since on CHERI
casting a pointer to a non-uintptr_t integer and back will result in
a pointer without the validity bit set. On dereference this then results
in a trap. With this patch I am able to successfully run the test suite
on CHERI-RISC-V without any failures.

Since C89 is still supported this patch introduces a JUINTPTR typedef
that can be used to avoid the C99 <stdint.h>.