Add new build option -DSUFFIX_LIBJPEG12_SYMBOLS=1
to be used with -DWITH_12BIT=1 to suffix symbols with _12, to avoid
clashes when linking an application against libjpeg 8 bit and libjpeg 12 bit.

This is an effort to revive an older effort done in 2009 by Frank
Warmerdam in the SourceForge repository post libjpeg-6b, and that was
used in libtiff and GDAL to be able to link against libjpeg in 8bit mode
and 12bit mode simultaneously.