jsimd_quantize_neon: Add #pragma unroll to help Clang code generation
The loop in jsimd_quantize_neon is only ever executed twice (based on
the DCT block size) and should be unrolled for AArch64 targets. GCC
does this by default but some Clang versions do not. This patch adds
an unroll pragma which is sufficient to get Clang to unroll the loop.
We do not use the unroll pragma for AArch32 targets as register
pressure would become a problem.

The effects of unrolling this loop are:
1) Elimination of the loop control-flow overhead.
2) Enabling the use of LDP/STP instructions working from a single
   base pointer, instead of using double the number of LDR/STR
   instructions - with each requiring address calculation.

Change-Id: I1b20aab272af9b66a512b832249fd842eb07bfee