for MSVC, generate a .rc file to embed version
This makes it so that when you right-click the DLL and
navigate to the Details tab, you find information
about the libjpeg-turbo version used to create the
emulated libjpeg library.

One CACHED parameter is added so that the user
can set their own CompanyName value, since that
field is traditionally used to describe the company
that actually built the file (as opposed to the
company that wrote the source, for example).