Fix CMake build for inclusion as a subproject
Introduces JPEG_ENABLE_TESTS option to control whether tests are added
with `add_tests`.  Defaults to OFF when built as a sub-project.

Prefixes all CMake macros and functions with `jpeg_`.  Note that the
CMake function/macro namespace is global.

Prefixes all targets with `jpeg_` if they did not already include
`jpeg` or similar in the name.   Note that the CMake target namespace
is global.

Adds `LANGUAGE ASM_NASM` property to x86 nasm sources.  Otherwise,
building fails when added as a sub-project of a larger project that
has both `ASM` and `ASM_NASM` languages enabled.