Fix buffer read overflow in encode_mcu_AC_first/encode_mcu_AC_refine
The pre-computed absolute values used in those functions were stored in
a JCOEF array.
The absolute value of -32768 can't be represented as a JCOEF value.
This led to -32768 being used as the absolute value of -32768 and,
further down the road, led to a buffer read overflow in emit_symbol.

This commit introduces a JABSCOEF type to store absolute values of
DCT coefficients and uses this new type to hold absolute values computed
by the preparations functions used in encode_mcu_AC_first/encode_mcu_AC_refine.