Test: Fix float test errors w/ Clang & fp-contract
The MD5 sums associated with FLOATTEST=fp-contract are appropriate for
GCC (tested v5 through v13) with -ffp-contract=fast, which is the
default when compiling for an architecture that has fused multiply-add
(FMA) instructions.  However, different MD5 sums are needed for Clang
(tested v5 through v14) with -ffp-contract=on, which is now the default
in Clang 14 when compiling for an architecture that has FMA
instructions.

Refer to #705, #709, #710