Build: Set FLOATTEST* by default for AArch64, PPC
Because of aa5efa3a1d90363b0884b87142b467c3c1487029, we can now
reliably determine the correct default value for FLOATTEST when using
Clang or GCC to build for AArch64 or PowerPC platforms.  (Testing
confirms that this is the case with GCC 5-13 and Clang 5-14 on
Ubuntu/AArch64, GCC 4 on CentOS 7/PPC, and GCC 8-10 and Clang 6-12 on
Ubuntu/PPCLE.)

Referring to 63bd71885f566dc65db9c9481fc1445f720451a7, it is not
possible to reliably determine the correct default value for FLOATTEST
when using other compilers or building for other architectures until
those compilers and architectures are thoroughly tested.  Thus, we no
longer attempt to set a default value except with tested configurations.
Other CPU architectures and compilers can be added on a case-by-case
basis as they are tested.