Test: fix with emulators that don't check cwd
While QEMU will run executables from the working directory, other
emulators may not.  A more reliable thing to do is to have the
emulator run with the full path to the executable.  The
add_test(NAME ... COMMAND ...) form handles emulation correctly
automatically as long as the first COMMAND argument is the name of a
target, so by switching to that, CMake will take care of setting up
emulation for free, and it will correctly give the emulator the full
path to the executable.