appveyor.yml: Require Java 8 or later
Public support for Java 6 ended in April 2013, and extended support for
it ended in October 2018.  Public support for Java 7 ended in July 2015,
and extended support for it ended in July 2022.  Technically speaking,
the only code that requires Java 8 at the moment is in TJUnitTest.java
(introduced in 79b8d65f0f86af77afc5979ecc104b1fbc97c82d.)  It is
possible to refactor that code to be compatible with Java 6, but there
are much better uses of our project's limited resources than supporting
long-dead platforms.