TJ: Fix OSS-Fuzz transform fuzzer ASan failure
(introduced by previous commit)

tjBufSize()-- and, by extension, tjTransform()-- cannot handle arbitrary
subsampling levels, so tjTransform() needs to check for that regardless
of whether lossless cropping will be performed.  This commit also
modifies the transform fuzzer itself so that it takes into account the
subsampling level of the destination image when computing the worst-case
buffer size for 90-degree rotation.