tjTransform(): Allow arb. subsamp w/ buf. realloc
We only need to care about the destination image's level of subsampling
if TJFLAG_NOREALLOC is set or lossless cropping will be performed.

This commit also causes tjTransform() to reject 4:1:1 source images if
TJFLAG_NOREALLOC is set or lossless cropping will be performed and if
the transform would generate a 4:4:1 destination image.  (This version
of the TurboJPEG API can't compute the worst-case JPEG buffer size for
4:4:1 images.)