GitHub: Don't test x32 ABI
The ubuntu-20.04 hosted runner image is going away on April 1, and newer
versions of Ubuntu can build but not run x32 binaries.  The x32 ABI
seems to be mostly dead, but we can still regression test x32 support
using a local Ubuntu 20.04 VM if necessary.  (That shouldn't be
necessary unless the x86-64 SIMD extensions change at some point in the
future.)