Build/CI: Officially support Windows/Arm
- Use Visual Studio 2022 for CI builds.
- Rename installers to clearly distinguish x86, x64, and Arm64 versions.

NOTE: The Windows/Arm installer uses the same internal name and install
directory as the Windows/x64 installer.  That is consistent with the
behavior of the Linux packages.  Because the installer places
turbojpeg.dll into C:\WINDOWS\System32, it would normally be impossible
to co-install the x64 and Arm64 versions anyhow.  (That can still be
accomplished.  You just have to use 7-Zip to extract one of the
installers into a non-default directory.)