Loongson: Don't autodetect MMI w/ -mloongson-mmi
As with other architectures, there is no need to autodetect the
availability of a particular SIMD instruction set if the compiler
targets an architecture that requires that instruction set.  If the code
is built with -mloongson-mmi, then the resulting binary will not work on
CPUs that lack MMI instructions, so libjpeg-turbo can safely assume that
MMI instructions are available.

More specifically, using /proc/cpuinfo to autodetect SIMD instruction
set availability is unreliable under QEMU, and QEMU is the only
mechanism by which I can test the correctness of the DSPr2 and MMI
SIMD extensions.